+++ examples/data-sources/meraki_appliance_dns_local_record/data-source.tf
@@ -1,0 +1,4 @@
+data "meraki_appliance_dns_local_record" "example" {
+  id              = "12345678"
+  organization_id = "123456"
+}