+++ examples/data-sources/meraki_appliance_dns_local_records/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_dns_local_records" "example" {
+  organization_id = "123456"
+}