+++ examples/resources/meraki_appliance_dns_local_record/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_dns_local_record.example "<organization_id>,<id>"